+++ core/src/components/jeep-stretchy-header/jeep-stretchy-header.spec.ts
@@ -1,0 +1,7 @@
+import { JeepStretchyHeader } from './jeep-stretchy-header';
+
+describe('jeep-stretchy-header', () => {
+  it('builds', () => {
+    expect(new JeepStretchyHeader()).toBeTruthy();
+  });
+});