+++ core/src/global/interfaces/stretchyheader.ts
@@ -1,0 +1,4 @@
+export interface StretchyHeaderToolbar {
+    contrastColor?: string;
+    color?:string;
+}