+++ pwa-components-viewme/src/components/app-stretchyheader-viewme/app-stretchyheader-viewme.spec.ts
@@ -1,0 +1,8 @@
+import { AppStretchyheaderViewme} from './app-stretchyheader-viewme';
+
+describe('app-stretchyheader-viewme', () => {
+  it('should build', () => {
+    expect(new AppStretchyheaderViewme()).toBeTruthy();
+  });
+  
+});