+++ pwa-components-viewme/src/components/app-stretchyheader/app-stretchyheader.spec.ts
@@ -1,0 +1,8 @@
+import { AppStretchyheader } from './app-stretchyheader';
+
+describe('app-stretchyheader', () => {
+  it('should build', () => {
+    expect(new AppStretchyheader()).toBeTruthy();
+  });
+
+});